chore: move tests to the separate workspace
