+++ .gitignore
@@ -8,6 +8,7 @@ node_modules
 
 # testing
 coverage
+cypress.env.json
 
 # next.js
 .next/
