+++ apps/lend/.gitignore
@@ -7,7 +7,6 @@
  
 # testing
 /coverage
-cypress.env.json
 
 # next.js
 /.next/
