+++ tests/cypress.config.ts
@@ -13,6 +13,6 @@ export default defineConfig({
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },
-    baseUrl: 'http://localhost:3003/#',
+    // baseUrl: 'http://localhost:3003/#',
   },
 })
