+++ tests/cypress/support/collateral.ts
