+++ tests/cypress/tools/Eip1193BridgeV6.ts
